Move Eula, InstanceOverview, Layout and NotFound component to pages
