+++ assets/js/pages/Eula/Eula.jsx
