+++ assets/js/pages/Eula/Eula.test.jsx
