+++ assets/js/pages/Eula/index.js
