+++ assets/js/pages/Layout/index.js
